Move cache to nodemodules node
